Change Nodes and started moving platform support
Changed nodes from [Tool] to [GlobalClass], This allows for better node organisation in AddNode. Swapped nodes from using Interfaces to Inheritance so common functionality can be shared and not rely on inheritance as a work around for AddCustomType() only being able to add to a base type and not other extended Nodes.